// projects/mb/src/lib/trading/indicators/statistics/standard-deviation/standard-deviation-params.interface.ts
import { BarComponent } from '../../../../data/entities/bar-component.enum';
import { QuoteComponent } from '../../../../data/entities/quote-component.enum';

/** Describes parameters to create an instance of the indicator. */
export interface StandardDeviationParams {
    /**
     * Length is the length (the number of time periods, ℓ) of the moving window to calculate the standard deviation.
     *
     * The value should be greater than 1.
     */
    length: number;

    /**
     * Unbiased indicates whether the estimate of the standard deviation is the unbiased sample standard deviation
     * or the population standard deviation.
     *
     * When in doubt, use the unbiased sample standard deviation (value is true).
     */
    unbiased: boolean;

    /**
     * A component of a bar to use when updating the indicator with a bar sample.
     *
     * If _undefined_, the bar component will have a default value and will not be shown in the indicator mnemonic.
     */
    barComponent?: BarComponent;

    /**
     * A component of a quote to use when updating the indicator with a quote sample.
     *
     * If _undefined_, the quote component will have a default value and will not be shown in the indicator mnemonic.
     */
    quoteComponent?: QuoteComponent;
}